Update README.md
Introduction
Online retail is a transactional data set which contains all the transactions occurring between 01/12/2010 and 09/12/2011 for a UK-based and registered non-store online retail. The company mainly sells unique all-occasion gifts. Many customers of the company are wholesalers.

We aim to segment the Customers based on RFM(Recency, Frequency, Monetary) so that the company can target its customers efficiently.

Recency — How recently did the customer purchase? It means the number of days since a customer made the last purchase.
Frequency — How often do they purchase in a given period. So we can understand this value as for how often or how many a customer used the product of a company. The bigger the value is, the more engaged the customers are.
Monetary — How much do they spend? It is the total amount of money a customer spent in a given period. Therefore big spenders will be differentiated with other customers such as MVP or VIP.